refactor: move code to lib + test existance of exported globals
