+++ schemata.js
@@ -1,67 +1,13 @@
 const isPrimitive = require('is-primitive')
 const clone = require('lodash.clonedeep')
 const SchemataArray = require('./lib/array')
+const castProperty = require('./lib/property-caster')
 const hasTag = require('./lib/has-tag')
 const isSchemata = require('./lib/is-schemata')
 const isSchemataArray = require('./lib/is-array')
 const getType = require('./lib/type-getter')
-const castArray = require('./lib/casters/array')
-const castBoolean = require('./lib/casters/boolean')
-const castDate = require('./lib/casters/date')
-const castNumber = require('./lib/casters/number')
-const castObject = require('./lib/casters/object')
-const castString = require('./lib/casters/string')
 const { validate, validateRecursive } = require('./lib/validate')
 const convertCamelcaseToHuman = require('./lib/camelcase-to-human-converter')
-/**
- * Casts a value to a given type.
- *
- * For booleans and integers; undefined, '', and null will all be cast to null
- * For array they will be converted to []
- * For object they will be converted to {}
- *
- * Throws error if type is undefined.
- *
- */
-const castProperty = (type, value, key, entityObject) => {
-  if (type === undefined) throw new Error('Missing type')
-
-  // First check whether the type of this property is
-  // a sub-schema, or an array of sub-schemas
-
-  const subSchema = getType(type, entityObject)
-  if (isSchemata(subSchema)) {
-    return value !== null ? subSchema.cast(value) : null
-  }
-
-  if (isSchemataArray(type)) {
-    if (!value) return null
-    if (!Array.isArray(value)) value = [value]
-    return value.map(v => type.arraySchema.cast(v))
-  }
-
-  // If the { type: x } is a primitive constructor, use
-  // cast the value based on which constructor is found
-
-  // JSHint doesn't like switch statements!
-  /* jshint maxcomplexity: 13 */
-  switch (type) {
-    case Boolean:
-      return castBoolean(value)
-    case Number:
-      return castNumber(value)
-    case String:
-      return castString(value)
-    case Object:
-      return castObject(value)
-    case Date:
-      return castDate(value)
-    case Array:
-      return castArray(value)
-    default:
-      return value
-  }
-}
 
 const createSchemata = ({ name, description, properties } = {}) => {
   if (name === undefined) throw new Error('name is required')
@@ -187,8 +133,6 @@ const createSchemata = ({ name, description, properties } = {}) => {
      * then only properties with that tag will remain.
      */
     stripUnknownProperties(entityObject, tag, ignoreTagForSubSchemas) {
-      /* jshint maxcomplexity: 10 */
-
       const newEntity = {}
 
       Object.keys(entityObject).forEach(key => {
